Revert structure change; split into modules; re-arrange packages
