+++ core/src/main/java/apdu4j/HexUtils.java
