+++ core/src/main/java/apdu4j/TagRemovedException.java
@@ -1,0 +1,13 @@
+package apdu4j;
+
+import java.io.IOException;
+
+public class TagRemovedException extends IOException {
+    public TagRemovedException(String message) {
+        super(message);
+    }
+
+    public TagRemovedException(String message, Throwable e) {
+        super(message, e);
+    }
+}
