+++ json-remote-emv/pom.xml
@@ -1,0 +1,31 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <parent>
+        <artifactId>apdu4j</artifactId>
+        <groupId>com.github.martinpaljak</groupId>
+        <version>18.05.01</version>
+    </parent>
+    <modelVersion>4.0.0</modelVersion>
+
+    <artifactId>json-remote-emv</artifactId>
+    <version>18.05.01</version>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.github.martinpaljak</groupId>
+            <!-- FIXME: this could be core-->
+            <artifactId>apdu4j-pcsc</artifactId>
+            <version>18.05.01</version>
+        </dependency>
+        <!-- Remote EMV JSON Reader -->
+        <dependency>
+            <groupId>com.googlecode.json-simple</groupId>
+            <artifactId>json-simple</artifactId>
+            <version>1.1.1</version>
+            <!-- TODO: jackson -->
+            <optional>true</optional>
+        </dependency>
+    </dependencies>
+</project>
