+++ json-remote-emv/src/main/java/apdu4j/remote/CmdlineRemoteTerminal.java
