+++ json-remote-emv/src/main/java/apdu4j/remote/RemoteTerminalThread.java
