+++ json-remote-emv/src/main/java/apdu4j/remote/SocketTransport.java
