+++ pcsc/pom.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+
+    <parent>
+        <groupId>com.github.martinpaljak</groupId>
+        <artifactId>apdu4j</artifactId>
+        <version>18.05.01</version>
+    </parent>
+
+    <groupId>com.github.martinpaljak</groupId>
+    <artifactId>apdu4j-pcsc</artifactId>
+    <version>18.05.01</version>
+
+    <dependencies>
+        <dependency>
+            <groupId>com.github.martinpaljak</groupId>
+            <artifactId>apdu4j-core</artifactId>
+            <version>18.05.01</version>
+        </dependency>
+        <!-- PC/SC access -->
+        <dependency>
+            <groupId>com.github.martinpaljak</groupId>
+            <artifactId>apdu4j-jnasmartcardio</artifactId>
+            <version>0.2.7+190425</version>
+        </dependency>
+    </dependencies>
+</project>
