--- src/main/java/apdu4j/SCard.java
@@ -1,6 +1,0 @@
-package apdu4j;
-
-public final class SCard {
-    public static final String SCARD_E_SHARING_VIOLATION = "SCARD_E_SHARING_VIOLATION";
-    public static final String SCARD_E_NO_READERS_AVAILABLE = "SCARD_E_NO_READERS_AVAILABLE";
-}
