+++ testing/src/main/java/apdu4j/testing/ParallelReactor.java
@@ -1,0 +1,4 @@
+package apdu4j.testing;
+
+public class ParallelReactor {
+}
